// src/modules/purchaser/components/create/side-bar/types.tsx
export type Profile = {
  fullName: string;
  personType: string;
  ocupation: number;
  ocupationDescription: string;
  country: string;
  countryDescription: string;
  savedRfc: string;
  curp: string;
  numId?: string;
  tradeName?: string;
  birthDateLegal?: string;
  register?: any;
  representative?: any;
};